move resource related classes to spark.resource
