+++ core/src/main/scala/org/apache/spark/executor/CoarseGrainedExecutorBackend.scala
@@ -27,14 +27,15 @@ import scala.util.{Failure, Success}
 import scala.util.control.NonFatal
 
 import org.json4s.DefaultFormats
-
 import org.apache.spark._
-import org.apache.spark.ResourceUtils._
+
+import org.apache.spark.resource.ResourceUtils._
 import org.apache.spark.TaskState.TaskState
 import org.apache.spark.deploy.SparkHadoopUtil
 import org.apache.spark.deploy.worker.WorkerWatcher
 import org.apache.spark.internal.Logging
 import org.apache.spark.internal.config._
+import org.apache.spark.resource.ResourceInformation
 import org.apache.spark.rpc._
 import org.apache.spark.scheduler.{ExecutorLossReason, TaskDescription}
 import org.apache.spark.scheduler.cluster.CoarseGrainedClusterMessages._
