+++ core/src/main/scala/org/apache/spark/scheduler/TaskSetManager.scala
@@ -27,9 +27,11 @@ import scala.math.max
 import scala.util.control.NonFatal
 
 import org.apache.spark._
+
 import org.apache.spark.TaskState.TaskState
 import org.apache.spark.internal.{config, Logging}
 import org.apache.spark.internal.config._
+import org.apache.spark.resource.ResourceInformation
 import org.apache.spark.scheduler.SchedulingMode._
 import org.apache.spark.util.{AccumulatorV2, Clock, LongAccumulator, SystemClock, Utils}
 import org.apache.spark.util.collection.MedianHeap
