+++ resource-managers/yarn/src/main/scala/org/apache/spark/deploy/yarn/YarnSparkHadoopUtil.scala
@@ -24,10 +24,11 @@ import scala.collection.mutable.{HashMap, ListBuffer}
 import org.apache.hadoop.yarn.api.ApplicationConstants
 import org.apache.hadoop.yarn.api.records.{ApplicationAccessType, ContainerId, Priority}
 import org.apache.hadoop.yarn.util.ConverterUtils
+import org.apache.spark.{SecurityManager, SparkConf}
 
-import org.apache.spark.{ResourceID, SecurityManager, SparkConf}
-import org.apache.spark.ResourceUtils._
+import org.apache.spark.resource.ResourceUtils._
 import org.apache.spark.launcher.YarnCommandBuilderUtils
+import org.apache.spark.resource.ResourceID
 import org.apache.spark.util.Utils
 
 object YarnSparkHadoopUtil {
